requirements.txt
google-generativeai>=0.7.0
python-dotenv>=1.0.1
pillow>=10.4.0
